+++ src/main/java/com/yeoljeong/tripmate/infrastructer/messaging/consumer/OrderEventListener.java
@@ -2,7 +2,6 @@ package com.yeoljeong.tripmate.infrastructer.messaging.consumer;
 
 import com.yeoljeong.tripmate.application.service.command.PlanInternalCommandService;
 import com.yeoljeong.tripmate.event.OrderCreatedEvent;
-import com.yeoljeong.tripmate.event.PaymentCompletedEvent;
 import com.yeoljeong.tripmate.event.enums.OrderTopic;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,7 +11,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @RequiredArgsConstructor
 @Component
-public class OrderCreatedEventListener {
+public class OrderEventListener {
 
   private final PlanInternalCommandService planInternalCommandService;
 
@@ -36,18 +35,4 @@ public class OrderCreatedEventListener {
     }
   }
 
-  public void paymentCompleted(PaymentCompletedEvent event) {
-    try {
-      if (event == null) {
-        log.warn("[plan-service] 결제 완료 이벤트가 null입니다.");
-        return;
-      }
-      planInternalCommandService.updateParticipantStatus(event);
-      log.info("[plan-service] 메시지 업데이트 처리 완료 : eventId={} ", event.eventHash());
-
-    } catch (Exception e) {
-      log.error("[plan-service] 메시지 업데이트 처리 실패 : {} ", e.getMessage(), e);
-      throw e;
-    }
-  }
 }
