+++ src/main/java/com/yeoljeong/tripmate/infrastructer/messaging/consumer/PaymentEventListener.java
@@ -1,0 +1,32 @@
+package com.yeoljeong.tripmate.infrastructer.messaging.consumer;
+
+import com.yeoljeong.tripmate.application.service.command.PlanInternalCommandService;
+import com.yeoljeong.tripmate.event.PaymentCompletedEvent;
+import com.yeoljeong.tripmate.event.enums.PaymentTopic;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@RequiredArgsConstructor
+@Component
+public class PaymentEventListener {
+  private final PlanInternalCommandService planInternalCommandService;
+
+  @KafkaListener(topics = PaymentTopic.PAYMENT_COMPLETED_TOPIC, groupId = "plan-group")
+  public void paymentCompleted(PaymentCompletedEvent event) {
+    try {
+      if (event == null) {
+        log.warn("[plan-service] 결제 완료 이벤트가 null입니다.");
+        return;
+      }
+      planInternalCommandService.updateParticipantStatus(event);
+      log.info("[plan-service] 메시지 업데이트 처리 완료 : eventId={} ", event.eventHash());
+
+    } catch (Exception e) {
+      log.error("[plan-service] 메시지 업데이트 처리 실패 : {} ", e.getMessage(), e);
+      throw e;
+    }
+  }
+}
